Work update
1. Installed material icons [material-icons](https://www.npmjs.com/package/material-icons)
2. Changed all old icons whit the new
3. Changed class name in `_components.scss`
4. Added logo as favicon

[Material Icons](https://marella.me/material-icons/demo/) - The list of icons to be used.

For adding new icons use the following
`<span class="material-icons">THE NAME OF THE ICON GOES HERE</span>`

To get the style correct, you will need to include `import 'material-icons/iconfont/material-icons.css';` in the script tag on the `page_name.svelte` where the icons will be used.

In the `_components.scss` file there is a `@mixin icon` section that was used to add style to icons before the change.
I removed `.icon-f` and `.icon-o` and replaced the class name with `.material-icons` but did not see any change in the style. Not sure why this is.

I also added the logo as a `favicon.ico`, For some reason it do not show up. Might work more on that.